+++ frontend/src/metabase/visualizations/components/ObjectDetail/ObjectDetail.tsx
@@ -16,12 +16,9 @@ import {
   useActionListQuery,
   useDatabaseListQuery,
 } from "metabase/common/hooks";
-import Button from "metabase/core/components/Button";
 import { NotFound } from "metabase/containers/ErrorPages";
-import EntityMenu from "metabase/components/EntityMenu";
 import LoadingSpinner from "metabase/components/LoadingSpinner";
 import Modal from "metabase/components/Modal";
-import { Flex } from "metabase/ui/components";
 
 import Tables from "metabase/entities/tables";
 import {
@@ -62,14 +59,12 @@ import {
   getSingleResultsRow,
 } from "./utils";
 import {
-  CloseButton,
   ErrorWrapper,
   ObjectDetailContainer,
-  ObjectDetailHeaderWrapper,
   ObjectDetailWrapperDiv,
-  ObjectIdLabel,
 } from "./ObjectDetail.styled";
 import { ObjectDetailBody } from "./ObjectDetailBody";
+import { ObjectDetailHeader } from "./ObjectDetailHeader";
 
 const mapStateToProps = (state: State, { data }: ObjectDetailProps) => {
   const isLoggedIn = !!getUser(state);
@@ -441,93 +436,6 @@ export function ObjectDetailView({
         />
       </Modal>
     </>
-  );
-}
-
-export interface ObjectDetailHeaderProps {
-  actionItems: {
-    title: string;
-    icon: string;
-    action: () => void;
-  }[];
-  canZoom: boolean;
-  objectName: string;
-  objectId: ObjectId | null | unknown;
-  canZoomPreviousRow: boolean;
-  canZoomNextRow?: boolean;
-  showControls?: boolean;
-  viewPreviousObjectDetail: () => void;
-  viewNextObjectDetail: () => void;
-  closeObjectDetail: () => void;
-}
-
-export function ObjectDetailHeader({
-  actionItems,
-  canZoom,
-  objectName,
-  objectId,
-  canZoomPreviousRow,
-  canZoomNextRow,
-  showControls = true,
-  viewPreviousObjectDetail,
-  viewNextObjectDetail,
-  closeObjectDetail,
-}: ObjectDetailHeaderProps): JSX.Element {
-  return (
-    <ObjectDetailHeaderWrapper className="Grid">
-      <div className="Grid-cell">
-        <h2 className="p3">
-          {objectName}
-          {objectId !== null && <ObjectIdLabel> {objectId}</ObjectIdLabel>}
-        </h2>
-      </div>
-
-      {showControls && (
-        <Flex align="center" gap="0.5rem" p="1rem">
-          {canZoom && (
-            <>
-              <Button
-                data-testid="view-previous-object-detail"
-                onlyIcon
-                borderless
-                disabled={!canZoomPreviousRow}
-                onClick={viewPreviousObjectDetail}
-                icon="chevronup"
-              />
-              <Button
-                data-testid="view-next-object-detail"
-                onlyIcon
-                borderless
-                disabled={!canZoomNextRow}
-                onClick={viewNextObjectDetail}
-                icon="chevrondown"
-              />
-            </>
-          )}
-
-          {actionItems.length > 0 && (
-            <EntityMenu
-              horizontalAttachments={["right", "left"]}
-              items={actionItems}
-              triggerIcon="ellipsis"
-              triggerProps={{
-                "data-testid": "actions-menu",
-              }}
-            />
-          )}
-
-          <CloseButton>
-            <Button
-              data-testid="object-detail-close-button"
-              onlyIcon
-              borderless
-              onClick={closeObjectDetail}
-              icon="close"
-            />
-          </CloseButton>
-        </Flex>
-      )}
-    </ObjectDetailHeaderWrapper>
   );
 }
